decomp: cast MakeString %d args to int — type-mangling fixes
bin/find-inlining-gaps cluster: MakeString<i>/MakeString<Ui> mismatch.
Target's source passed explicit (int) casts where ours emitted
MakeString<unsigned int> / MakeString<unsigned short>.

- ChordShapeGenerator BuildContourCap/BuildEndCap "too few faces"
  MILO_LOG: newsize was unsigned int -> cast to int (target wants i)
- StoreMainPanel ParseConfigData "%d" marquee[4]: unsigned short ->
  (int) cast (target wants i)

BuildContourCap 98.8 -> 98.9% (1 bl mismatch dropped; FPR cascade
remains). StoreMainPanel still at 99.1% AtLimit (one fewer diff_arg,
r4↔r5 swap dominates). The verts MILO_LOGs correctly stayed as
unsigned int — target wants <Ui> for those sites.

+++ src/band3/meta_band/StoreMainPanel.cpp
@@ -150,7 +150,7 @@ void StoreMainPanel::ParseConfigData() {
         entry.mText1 = TheStoreMetadata.GetString(marquee[2]);
         entry.mText2 = TheStoreMetadata.GetString(marquee[3]);
         entry.mText3 = TheStoreMetadata.GetString(marquee[0]);
-        entry.mText4 = MakeString("%d", marquee[4]);
+        entry.mText4 = MakeString("%d", (int)marquee[4]);
         EnsureArtLoader(entry.mStrName);
         mNewReleaseList.push_back(entry);
         mCoverArtTexs.push_back(Hmx::Object::New<RndTex>());
